+++ frontend/src/hooks/index.js
@@ -4,3 +4,5 @@
 export { usePermission } from './usePermission'
 export { useDebounce, useDebouncedCallback } from './useDebounce'
 export { useDeleteHandler, useFormData, useModals, usePagination, useAsyncData } from './useCommon'
+export { useWebSocket, EventType, ConnectionState } from './useWebSocket'
+export { useAutoRefresh, CertificateEvents, CAEvents, UserEvents, GroupEvents, CRLEvents, AllPKIEvents } from './useAutoRefresh'